requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=1.4.41
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
pytest>=7.4.0
httpx>=0.24.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6 